refactor & better errors
