+++ gcalbot/gcalbot/handler.go
@@ -1,7 +1,6 @@
 package gcalbot
 
 import (
-	"fmt"
 	"strings"
 
 	"github.com/kballard/go-shellquote"
@@ -88,101 +87,4 @@ func (h *Handler) HandleCommand(msg chat1.MsgSummary) error {
 		// TODO(marcel): send user an error message
 		return nil
 	}
-}
-
-func (h *Handler) accountsListHandler(msg chat1.MsgSummary) error {
-	username := msg.Sender.Username
-	accounts, err := h.db.GetAccountsForUser(username)
-	if err != nil {
-		h.ChatDebug(msg.ConvID, "error fetching accounts from database %q", err)
-		return nil
-	}
-	accountListMessage := "Here are your connected accounts:"
-	for index := range accounts {
-		accountListMessage += fmt.Sprintf("\n• %s", accounts[index])
-	}
-	_, err = h.kbc.SendMessageByConvID(msg.ConvID, accountListMessage)
-	return err
-}
-
-func (h *Handler) accountsConnectHandler(cmd string, msg chat1.MsgSummary) error {
-	toks, err := shellquote.Split(cmd)
-	if err != nil {
-		// TODO(marcel): possibly a better error message
-		h.ChatDebug(msg.ConvID, "error splitting command: %s", err)
-		return nil
-	}
-	args := toks[3:]
-	if len(args) != 1 {
-		// TODO(marcel): possibly a better error message
-		h.ChatDebug(msg.ConvID, "bad args for accounts connect: %s", args)
-		return nil
-	}
-
-	username := msg.Sender.Username
-	accountNickname := args[0]
-
-	exists, err := h.db.ExistsAccountForUser(username, accountNickname)
-	if err != nil {
-		h.ChatDebug(msg.ConvID, "error checking account: %s", err.Error())
-		return nil
-	}
-	if exists {
-		_, err = h.kbc.SendMessageByConvID(msg.ConvID,
-			"An account connection with the nickname '%s' already exists.", accountNickname)
-		return err
-	}
-
-	// TODO(marcel): is this the best method? adding more columns to the db complicates the existing oauth abstractions
-	identifier := fmt.Sprintf("%s:%s", username, accountNickname)
-	// TODO(marcel): this is very hacky
-	authMessageTemplate := fmt.Sprintf("Visit %s to connect a Google account as '%s'.", "%s", accountNickname)
-
-	_, err = base.GetOAuthClient(identifier, msg, h.kbc, h.requests, h.config, h.db,
-		authMessageTemplate, base.GetOAuthOpts{OAuthOfflineAccessType: true})
-	return err
-}
-
-func (h *Handler) accountsDisconnectHandler(cmd string, msg chat1.MsgSummary) error {
-	toks, err := shellquote.Split(cmd)
-	if err != nil {
-		// TODO(marcel): possibly a better error message
-		h.ChatDebug(msg.ConvID, "error splitting command: %s", err)
-		return nil
-	}
-	args := toks[3:]
-	if len(args) != 1 {
-		// TODO(marcel): possibly a better error message
-		h.ChatDebug(msg.ConvID, "bad args for accounts connect: %s", args)
-		return nil
-	}
-
-	username := msg.Sender.Username
-	accountNickname := args[0]
-
-	exists, err := h.db.ExistsAccountForUser(username, accountNickname)
-	if err != nil {
-		h.ChatDebug(msg.ConvID, "error checking account: %s", err.Error())
-		return nil
-	}
-	if !exists {
-		_, err = h.kbc.SendMessageByConvID(msg.ConvID,
-			"No account connection with the nickname '%s' exists.", accountNickname)
-		return err
-	}
-
-	// TODO(marcel): is this the best method? adding more columns to the db complicates the existing oauth abstractions
-	identifier := fmt.Sprintf("%s:%s", username, accountNickname)
-	// TODO(marcel): wrap these into one transcation
-	err = h.db.DeleteToken(identifier)
-	if err != nil {
-		return err
-	}
-	err = h.db.DeleteAccountForUser(username, accountNickname)
-	if err != nil {
-		return err
-	}
-
-	_, err = h.kbc.SendMessageByConvID(msg.ConvID, "Account '%s' has been disconnected.", accountNickname)
-	return err
 }
